+++ Models/Usuario.cs
@@ -6,7 +6,7 @@ namespace TesteSeniors.Models
     {
         // Utilizado o identity User, por ja possuir como base algumas validacoes e regras de negocio
         public string Nome { get; set; }
-        public int CPF { get; set; }
+        public long CPF { get; set; }
         public string UF { get; set; }
         public DateTime DataNascimento { get; set; }
         public Usuario(): base() { }
